+++ CHANGES.md
@@ -1,1 +1,2 @@
 - 2016/04/28 -- Initial project
+- 2016/12/20 -- Support optimize all response